requirements.txt
requests
Flask
flask_login
pyopenssl
simple_salesforce
python-dotenv